Removed the service from program main and use lazy singleton.
