+++ src/BlazorAdmin/Pages/Login.razor
@@ -4,7 +4,7 @@
 @inject AuthService Auth
 @inject NavigationManager NavigationManager
 
-@inherits BlazorAdmin.Helper.BlazorComponent
+@inherits BlazorAdmin.Helpers.BlazorComponent
 
 <div class="container">
     @if (ShowLoginFailed)
