+++ src/BlazorAdmin/Pages/Logout.razor
@@ -3,7 +3,7 @@
 
 @inject AuthService AuthService
 @inject NavigationManager NavigationManager
-@inherits BlazorAdmin.Helper.BlazorComponent
+@inherits BlazorAdmin.Helpers.BlazorComponent
 
 
 @code {
