--- src/BlazorAdmin/Services/RefreshService.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace BlazorAdmin.Services
-{
-    public class RefreshService
-    {
-        public event Action RefreshRequested;
-        public void CallRequestRefresh()
-        {
-            RefreshRequested?.Invoke();
-        }
-    }
-}
