+++ src/BlazorAdmin/Shared/MainLayout.razor
@@ -1,7 +1,7 @@
 ﻿@using BlazorAdmin.Services
 
 @inject AuthService Auth
-@inherits BlazorAdmin.Helper.BlazorLayoutComponent
+@inherits BlazorAdmin.Helpers.BlazorLayoutComponent
 
 <div class="sidebar">
     <NavMenu />
