+++ src/BlazorAdmin/Shared/NavMenu.razor
@@ -1,5 +1,5 @@
 ﻿@inject AuthService Auth
-@inherits BlazorAdmin.Helper.BlazorComponent
+@inherits BlazorAdmin.Helpers.BlazorComponent
 
 @using BlazorAdmin.Services
 <div class="top-row pl-4 navbar navbar-dark">
